Model and Name change WIP
This update includes the first phase of a major model and name  change update.

Note that all occurances of "Arkane" have been replaced with "BGSDK" this is on Arkane's request as they are planning a possible name change as well.

In addition to the data model change the Unity editor extensions have been updated to use client secret for log in. This removes the need for a dev to log in through Unity.

We will follow up with additional updates in the coming week as we solidify the new architecture.